Implement tut05
Did not implement tutorial DepthFighting, as it is not covered in the book thus far, and it is fairly complex.

Added onKeyboard to GLTest, and set a callback for it.

Some reorganization

TODO: fix usage of glDrawElementsBaseVertex and depth buffer to be consistent with book ordering. The ordering got screwed up at some point.